+++ mbsimgui/mbsimgui/dialogs.cc
@@ -124,7 +124,7 @@ namespace MBSimGUI {
       mw->eval->convertIndex(value, true);
       count[idx]=mw->eval->cast<int>(value)-1;
     }
-    auto levels = mw->updateParameters(embeditem, nullptr, count);
+    auto levels = mw->updateParameters(embeditem, nullptr, false, count);
     count.resize(levels.size(), 0);
 
     auto var=widget->getEvalMat();