+++ mbsimgui/mbsimgui/treeitemdata.h
@@ -51,7 +51,6 @@ namespace MBSimGUI {
       void addTreeItemData(TreeItemData *treeItemData_) { treeItemData.push_back(treeItemData_); }
       void setModelIndex(const QModelIndex &index_) { index = index_; }
       const QModelIndex& getModelIndex() const { return index; }
-      void saveIcon() { orgIcon = icon; }
     protected:
       std::vector<TreeItemData*> treeItemData;
       QModelIndex index;